add subclass RpcTransactionStatusHandlerTest
