+++ platform/src/main/java/org/stellar/anchor/platform/rpc/NotifyCustomerInfoUpdatedHandler.java
@@ -24,7 +24,7 @@ import org.stellar.anchor.sep31.Sep31TransactionStore;
 import org.stellar.anchor.sep6.Sep6TransactionStore;
 
 public class NotifyCustomerInfoUpdatedHandler
-    extends RpcMethodHandler<NotifyCustomerInfoUpdatedRequest> {
+    extends RpcTransactionStatusHandler<NotifyCustomerInfoUpdatedRequest> {
 
   public NotifyCustomerInfoUpdatedHandler(
       Sep6TransactionStore txn6Store,
