+++ platform/src/main/java/org/stellar/anchor/platform/rpc/NotifyInteractiveFlowCompletedHandler.java
@@ -28,7 +28,7 @@ import org.stellar.anchor.sep31.Sep31TransactionStore;
 import org.stellar.anchor.sep6.Sep6TransactionStore;
 
 public class NotifyInteractiveFlowCompletedHandler
-    extends RpcMethodHandler<NotifyInteractiveFlowCompletedRequest> {
+    extends RpcTransactionStatusHandler<NotifyInteractiveFlowCompletedRequest> {
 
   public NotifyInteractiveFlowCompletedHandler(
       Sep6TransactionStore txn6Store,
