+++ platform/src/main/java/org/stellar/anchor/platform/rpc/NotifyOffchainFundsAvailableHandler.java
@@ -26,7 +26,7 @@ import org.stellar.anchor.sep31.Sep31TransactionStore;
 import org.stellar.anchor.sep6.Sep6TransactionStore;
 
 public class NotifyOffchainFundsAvailableHandler
-    extends RpcMethodHandler<NotifyOffchainFundsAvailableRequest> {
+    extends RpcTransactionStatusHandler<NotifyOffchainFundsAvailableRequest> {
 
   public NotifyOffchainFundsAvailableHandler(
       Sep6TransactionStore txn6Store,
