+++ platform/src/main/java/org/stellar/anchor/platform/rpc/NotifyOffchainFundsSentHandler.java
@@ -32,7 +32,7 @@ import org.stellar.anchor.sep31.Sep31TransactionStore;
 import org.stellar.anchor.sep6.Sep6TransactionStore;
 
 public class NotifyOffchainFundsSentHandler
-    extends RpcMethodHandler<NotifyOffchainFundsSentRequest> {
+    extends RpcTransactionStatusHandler<NotifyOffchainFundsSentRequest> {
 
   public NotifyOffchainFundsSentHandler(
       Sep6TransactionStore txn6Store,
