+++ platform/src/main/java/org/stellar/anchor/platform/rpc/NotifyRefundPendingHandler.java
@@ -42,7 +42,8 @@ import org.stellar.anchor.sep24.Sep24TransactionStore;
 import org.stellar.anchor.sep31.Sep31TransactionStore;
 import org.stellar.anchor.sep6.Sep6TransactionStore;
 
-public class NotifyRefundPendingHandler extends RpcMethodHandler<NotifyRefundPendingRequest> {
+public class NotifyRefundPendingHandler
+    extends RpcTransactionStatusHandler<NotifyRefundPendingRequest> {
 
   public NotifyRefundPendingHandler(
       Sep6TransactionStore txn6Store,
