+++ platform/src/main/java/org/stellar/anchor/platform/rpc/NotifyRefundSentHandler.java
@@ -41,7 +41,7 @@ import org.stellar.anchor.sep31.Sep31Refunds;
 import org.stellar.anchor.sep31.Sep31TransactionStore;
 import org.stellar.anchor.sep6.Sep6TransactionStore;
 
-public class NotifyRefundSentHandler extends RpcMethodHandler<NotifyRefundSentRequest> {
+public class NotifyRefundSentHandler extends RpcTransactionStatusHandler<NotifyRefundSentRequest> {
 
   public NotifyRefundSentHandler(
       Sep6TransactionStore txn6Store,
