+++ platform/src/main/java/org/stellar/anchor/platform/rpc/NotifyTransactionErrorHandler.java
@@ -24,7 +24,8 @@ import org.stellar.anchor.sep24.Sep24TransactionStore;
 import org.stellar.anchor.sep31.Sep31TransactionStore;
 import org.stellar.anchor.sep6.Sep6TransactionStore;
 
-public class NotifyTransactionErrorHandler extends RpcMethodHandler<NotifyTransactionErrorRequest> {
+public class NotifyTransactionErrorHandler
+    extends RpcTransactionStatusHandler<NotifyTransactionErrorRequest> {
 
   private final JdbcTransactionPendingTrustRepo transactionPendingTrustRepo;
