+++ platform/src/main/java/org/stellar/anchor/platform/rpc/NotifyTransactionOnHoldHandler.java
@@ -28,7 +28,7 @@ import org.stellar.anchor.sep31.Sep31TransactionStore;
 import org.stellar.anchor.sep6.Sep6TransactionStore;
 
 public class NotifyTransactionOnHoldHandler
-    extends RpcMethodHandler<NotifyTransactionOnHoldRequest> {
+    extends RpcTransactionStatusHandler<NotifyTransactionOnHoldRequest> {
 
   public NotifyTransactionOnHoldHandler(
       Sep6TransactionStore txn6Store,
