+++ platform/src/main/java/org/stellar/anchor/platform/rpc/RequestOffchainFundsHandler.java
@@ -29,7 +29,8 @@ import org.stellar.anchor.sep24.Sep24TransactionStore;
 import org.stellar.anchor.sep31.Sep31TransactionStore;
 import org.stellar.anchor.sep6.Sep6TransactionStore;
 
-public class RequestOffchainFundsHandler extends RpcMethodHandler<RequestOffchainFundsRequest> {
+public class RequestOffchainFundsHandler
+    extends RpcTransactionStatusHandler<RequestOffchainFundsRequest> {
 
   public RequestOffchainFundsHandler(
       Sep6TransactionStore txn6Store,
