+++ platform/src/test/kotlin/org/stellar/anchor/platform/rpc/GetTransactionHandlerTest.kt
@@ -10,8 +10,8 @@ import org.junit.jupiter.api.assertThrows
 import org.skyscreamer.jsonassert.JSONAssert
 import org.skyscreamer.jsonassert.JSONCompareMode
 import org.stellar.anchor.api.event.AnchorEvent
-import org.stellar.anchor.api.exception.rpc.InvalidParamsException
-import org.stellar.anchor.api.exception.rpc.InvalidRequestException
+import org.stellar.anchor.api.exception.BadRequestException
+import org.stellar.anchor.api.exception.NotFoundException
 import org.stellar.anchor.api.platform.GetTransactionResponse
 import org.stellar.anchor.api.platform.PlatformTransactionData
 import org.stellar.anchor.api.rpc.method.GetTransactionRpcRequest
@@ -20,12 +20,16 @@ import org.stellar.anchor.api.shared.Amount
 import org.stellar.anchor.api.shared.Customers
 import org.stellar.anchor.api.shared.StellarId
 import org.stellar.anchor.asset.AssetService
+import org.stellar.anchor.config.CustodyConfig
+import org.stellar.anchor.custody.CustodyService
 import org.stellar.anchor.event.EventService
-import org.stellar.anchor.metrics.MetricsService
 import org.stellar.anchor.platform.data.JdbcSep31Transaction
-import org.stellar.anchor.platform.validator.RequestValidator
+import org.stellar.anchor.platform.service.TransactionService
+import org.stellar.anchor.sep24.Sep24DepositInfoGenerator
 import org.stellar.anchor.sep24.Sep24TransactionStore
 import org.stellar.anchor.sep31.Sep31TransactionStore
+import org.stellar.anchor.sep38.Sep38QuoteStore
+import org.stellar.anchor.sep6.Sep6DepositInfoGenerator
 import org.stellar.anchor.sep6.Sep6TransactionStore
 import org.stellar.anchor.util.GsonUtils
 
@@ -35,25 +39,21 @@ class GetTransactionHandlerTest {
     private const val TX_ID = "testId"
     private const val INVALID_ID = "invalidId"
     private const val EMPTY_ID_ERROR_MESSAGE = "transaction id cannot be empty"
-    private const val TRANSACTION_NOT_FOUND = "Transaction with id[invalidId] is not found"
+    private const val TRANSACTION_NOT_FOUND = "transaction (id=invalidId) is not found"
   }
 
+  @MockK(relaxed = true) private lateinit var sep38QuoteStore: Sep38QuoteStore
   @MockK(relaxed = true) private lateinit var txn6Store: Sep6TransactionStore
-
   @MockK(relaxed = true) private lateinit var txn24Store: Sep24TransactionStore
-
   @MockK(relaxed = true) private lateinit var txn31Store: Sep31TransactionStore
-
-  @MockK(relaxed = true) private lateinit var requestValidator: RequestValidator
-
   @MockK(relaxed = true) private lateinit var assetService: AssetService
-
   @MockK(relaxed = true) private lateinit var eventService: EventService
-
-  @MockK(relaxed = true) private lateinit var metricsService: MetricsService
-
   @MockK(relaxed = true) private lateinit var eventSession: EventService.Session
-
+  @MockK(relaxed = true) private lateinit var sep6DepositInfoGenerator: Sep6DepositInfoGenerator
+  @MockK(relaxed = true) private lateinit var sep24DepositInfoGenerator: Sep24DepositInfoGenerator
+  @MockK(relaxed = true) private lateinit var custodyService: CustodyService
+  @MockK(relaxed = true) private lateinit var custodyConfig: CustodyConfig
+  private lateinit var transactionService: TransactionService
   private lateinit var handler: GetTransactionHandler
 
   @BeforeEach
@@ -61,16 +61,20 @@ class GetTransactionHandlerTest {
     MockKAnnotations.init(this, relaxUnitFun = true)
     every { eventService.createSession(any(), EventService.EventQueue.TRANSACTION) } returns
       eventSession
-    this.handler =
-      GetTransactionHandler(
+    transactionService =
+      TransactionService(
         txn6Store,
         txn24Store,
         txn31Store,
-        requestValidator,
+        sep38QuoteStore,
         assetService,
         eventService,
-        metricsService
+        sep6DepositInfoGenerator,
+        sep24DepositInfoGenerator,
+        custodyService,
+        custodyConfig
       )
+    handler = GetTransactionHandler(transactionService)
   }
 
   @Test
@@ -111,7 +115,7 @@ class GetTransactionHandlerTest {
   @Test
   fun test_handle_emptyId() {
     val request = GetTransactionRpcRequest.builder().transactionId("").build()
-    val ex = assertThrows<InvalidParamsException> { handler.handle(request) }
+    val ex = assertThrows<BadRequestException> { handler.handle(request) }
     assertEquals(EMPTY_ID_ERROR_MESSAGE, ex.message?.trimIndent())
   }
 
@@ -121,7 +125,7 @@ class GetTransactionHandlerTest {
     every { txn6Store.findByTransactionId(any()) } returns null
     every { txn24Store.findByTransactionId(any()) } returns null
     every { txn31Store.findByTransactionId(INVALID_ID) } returns null
-    val ex = assertThrows<InvalidRequestException> { handler.handle(request) }
+    val ex = assertThrows<NotFoundException> { handler.handle(request) }
     assertEquals(TRANSACTION_NOT_FOUND, ex.message?.trimIndent())
   }
 }
