+++ platform/src/test/kotlin/org/stellar/anchor/platform/rpc/GetTransactionsHandlerTest.kt
@@ -13,13 +13,17 @@ import org.stellar.anchor.api.platform.TransactionsSeps
 import org.stellar.anchor.api.rpc.method.GetTransactionsRpcRequest
 import org.stellar.anchor.api.sep.SepTransactionStatus
 import org.stellar.anchor.asset.AssetService
+import org.stellar.anchor.config.CustodyConfig
+import org.stellar.anchor.custody.CustodyService
 import org.stellar.anchor.event.EventService
-import org.stellar.anchor.metrics.MetricsService
 import org.stellar.anchor.platform.data.JdbcSep31Transaction
+import org.stellar.anchor.platform.service.TransactionService
 import org.stellar.anchor.platform.utils.PlatformTransactionHelper
-import org.stellar.anchor.platform.validator.RequestValidator
+import org.stellar.anchor.sep24.Sep24DepositInfoGenerator
 import org.stellar.anchor.sep24.Sep24TransactionStore
 import org.stellar.anchor.sep31.Sep31TransactionStore
+import org.stellar.anchor.sep38.Sep38QuoteStore
+import org.stellar.anchor.sep6.Sep6DepositInfoGenerator
 import org.stellar.anchor.sep6.Sep6TransactionStore
 import org.stellar.anchor.util.GsonUtils
 
@@ -28,21 +32,18 @@ class GetTransactionsHandlerTest {
     private val gson = GsonUtils.getInstance()
   }
 
+  @MockK(relaxed = true) private lateinit var sep38QuoteStore: Sep38QuoteStore
   @MockK(relaxed = true) private lateinit var txn6Store: Sep6TransactionStore
-
   @MockK(relaxed = true) private lateinit var txn24Store: Sep24TransactionStore
-
   @MockK(relaxed = true) private lateinit var txn31Store: Sep31TransactionStore
-
-  @MockK(relaxed = true) private lateinit var requestValidator: RequestValidator
-
   @MockK(relaxed = true) private lateinit var assetService: AssetService
-
   @MockK(relaxed = true) private lateinit var eventService: EventService
-
-  @MockK(relaxed = true) private lateinit var metricsService: MetricsService
-
   @MockK(relaxed = true) private lateinit var eventSession: EventService.Session
+  @MockK(relaxed = true) private lateinit var sep6DepositInfoGenerator: Sep6DepositInfoGenerator
+  @MockK(relaxed = true) private lateinit var sep24DepositInfoGenerator: Sep24DepositInfoGenerator
+  @MockK(relaxed = true) private lateinit var custodyService: CustodyService
+  @MockK(relaxed = true) private lateinit var custodyConfig: CustodyConfig
+  private lateinit var transactionService: TransactionService
 
   private lateinit var handler: GetTransactionsHandler
 
@@ -51,16 +52,20 @@ class GetTransactionsHandlerTest {
     MockKAnnotations.init(this, relaxUnitFun = true)
     every { eventService.createSession(any(), EventService.EventQueue.TRANSACTION) } returns
       eventSession
-    this.handler =
-      GetTransactionsHandler(
+    transactionService =
+      TransactionService(
         txn6Store,
         txn24Store,
         txn31Store,
-        requestValidator,
+        sep38QuoteStore,
         assetService,
         eventService,
-        metricsService
+        sep6DepositInfoGenerator,
+        sep24DepositInfoGenerator,
+        custodyService,
+        custodyConfig
       )
+    this.handler = GetTransactionsHandler(transactionService)
   }
 
   @Test
